// src/app/core/models/employeeinfo.model.ts
// Raw shape returned by GET api/EmployeeInfo/GetEmployeeInfo
export interface EmpInfo {
  empid: string;
  emp_Code: string;
  salutation: string;
  salutation_Id: number;
  name: string;
  mName: string;
  lName: string;
  gen: string;
  doj: string;
  dob: string;
  dow: string;
  marital: number;
  mobile: string;
  mobile2: string;
  email: string;
  personal_Email: string;
  biometric_Id: string;
  others1: string;
  others2: string;
  others3: string;
  others4: string;
  others5: string;
  bonuspayterm: string;
  doad: string;
  fatherN: string;
  husbandName: string;
  division_Name: string;
  dep_Name: string;
  designation_Name: string;
  divid: string;
  depart_Id: string;
  desig_Id: string;
  city_Id: number;
  bloodGrp: string;
  adhr: string;
  bank: string;
  bankAc: string;
  branchName: string;
  pan: string;
  ifc: string;
  micr: string;
  addr: string;
  addr1: string;
  pincode: string;
  addr3: string;
  addr4: string;
  pincode1: string;
  state: number;
  state1: number;
  basic_details: number;
  personal_details: number;
  residential_details: number;
  family_details: number;
  education_details: number;
  working_details: number;
  esI_enable: number;
  epF_enable: number;
  pT_enable: number;
  tds: number;
  portal_access: number;
  lwF_status: number;
  pension_Scheme: number;
  higher_Wages: number;
  memberId: string;
  ismemberIdNeeded: string;
  nationality: string;
  nameAsperBankDetails: string;
  isPysicalHandicap: string;
  pysicalHandicap: string;
}

// Mapped/flattened shape the component actually consumes
export interface EmployeeInfo {
  id: string;
  empCode: string;
  salutation: string;
  firstName: string;
  lastName: string;
  fullName: string;
  gender: string;
  dob: string;
  doj: string;
  marital: number;
  mobile: string;
  mobile2: string;
  email: string;
  divisionName: string;
  departmentName: string;
  designationName: string;
  cityId: number;
  bloodGroup: string;
  aadhaar: string;
  bank: string;
  bankAc: string;
  branchName: string;
  pan: string;
  ifsc: string;
  addr1: string;
  addr2: string;
  pincode: string;
  permAddr1: string;
  permAddr2: string;
  permPincode: string;
}